requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aim==3.17.5
aim-ui==3.17.5
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.2
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
base58==2.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.37.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.9
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.6
clip==0.2.0
clip-by-openai==1.1
cmake==3.25.0
comm==0.1.4
confection==0.1.1
contexttimer==0.3.3
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
-e git+https://github.com/KaiyangZhou/Dassl.pytorch.git@c61a1b570ac6333bd50fb5ae06aea59002fb20bb#egg=dassl
datasets==2.15.0
debugpy==1.7.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.15.0
dill==0.3.7
distlib==0.3.7
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.4.1
exceptiongroup==1.1.2
executing==1.2.0
fairscale==0.4.4
fastapi==0.101.0
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.24.1
gradio_client==0.0.8
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.19.4
identify==2.5.27
idna==3.4
imagecorruptions==1.1.2
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kaggle==1.5.16
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
linkify-it-py==2.0.2
lit==15.0.7
littleutils==0.2.2
llvmlite==0.40.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.0
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.3
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ogb==1.3.6
omegaconf==2.3.0
open-clip-torch==2.23.0
openai==0.27.0
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.7
outdated==0.2.2
overrides==7.4.0
packaging==23.0
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.2
peft==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
plotly==5.16.1
portalocker==2.7.0
pre-commit==3.4.0
preshed==3.0.8
progress==1.6
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==13.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytils==0.4.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
RestrictedPython==6.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
segment-analytics-python==2.2.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shtab==1.6.5
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.8
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tools==0.1.9
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.9.0
transformers==4.28.0
triton==2.0.0
trl==0.5.0
typer==0.7.0
typing_extensions==4.7.1
tyro==0.6.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.24.4
wandb==0.15.8
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.6
wilds==2.0.0
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.8.2
zipp==3.14.0
